+++ src/gocept/selenium/tests/test_wd_selenese.py
@@ -17,6 +17,7 @@ from gocept.selenium.screenshot import ScreenshotSizeMismatchError
 from gocept.selenium.wd_selenese import LOCATOR_JS, LOCATOR_JQUERY
 from gocept.selenium.wd_selenese import selenese_pattern_equals as match
 from gocept.selenium.wd_selenese import split_locator, split_option_locator
+from selenium.common.exceptions import ElementClickInterceptedException
 from selenium.common.exceptions import NoSuchElementException
 from selenium.common.exceptions import StaleElementReferenceException
 from selenium.webdriver.common.by import By
@@ -243,6 +244,16 @@ class AssertionTest(AssertionTests,
             self.selenium.waitForVisible('css=.foo.bar')
         self.assertEllipsis(
             'Timed out after 1.0 s. ...Unable to locate element...',
+            e.exception.msg)
+
+    def test_wait_for_retries_when_element_was_intercepted_in_click(self):
+        """It retries clicking when an element is intercepted by another."""
+        self.selenium.setTimeout(1000)
+        self.selenium.open('/intercepted.html')
+        with self.assertRaises(ElementClickInterceptedException) as e:
+            self.selenium.click('css=.intercepted')
+        self.assertEllipsis(
+            'Timed out after 1.0 s. Element ... not clickable ... obscures it',
             e.exception.msg)
 
     def test_wd_selense__Selense__selectParentFrame__1(self):